Move ZeroOneInflatedBeta to mixture module; use safe-value pattern in logp/logcdf
