+++ docs/source/api/distributions/continuous.rst
@@ -41,4 +41,3 @@ Continuous
    VonMises
    Wald
    Weibull
-   ZeroOneInflatedBeta
