+++ docs/source/api/distributions/mixture.rst
@@ -12,6 +12,7 @@ Mixture
    ZeroInflatedBinomial
    ZeroInflatedNegativeBinomial
    ZeroInflatedPoisson
+   ZeroOneInflatedBeta
    HurdlePoisson
    HurdleNegativeBinomial
    HurdleGamma
