+++ tests/distributions/test_mixture.py
@@ -25,6 +25,8 @@ from pytensor.tensor import TensorVariable
 from pytensor.tensor.random.op import RandomVariable
 from scipy.special import logsumexp
 
+import pymc as pm
+
 from pymc.distributions import (
     Beta,
     Categorical,
@@ -71,6 +73,7 @@ from pymc.sampling.forward import (
 from pymc.sampling.mcmc import sample
 from pymc.step_methods import Metropolis
 from pymc.testing import (
+    BaseTestDistributionRandom,
     Domain,
     Nat,
     NatSmall,
@@ -1752,3 +1755,211 @@ class TestHurdleDistributions:
         dlogp_val = dlogp_fn(ip)
 
         assert not np.any(np.isnan(dlogp_val)), f"dlogp contains NaN: {dlogp_val}"
+
+
+class TestZeroOneInflatedBeta(BaseTestDistributionRandom):
+    pymc_dist = pm.ZeroOneInflatedBeta
+    pymc_dist_params = {"zoi": 0.3, "coi": 0.4, "mu": 0.5, "kappa": 10.0}
+    expected_rv_op_params = {"zoi": 0.3, "coi": 0.4, "a": 5.0, "b": 5.0}
+    size = 15
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_rv_size",
+    ]
+
+
+class TestZeroOneInflatedBetaLogp:
+    def test_logp_at_zero(self):
+        zoi, coi, mu, kappa = 0.3, 0.4, 0.4, 10.0
+        expected = np.log(zoi) + np.log1p(-coi)
+        with pm.Model():
+            dist = pm.ZeroOneInflatedBeta.dist(zoi=zoi, coi=coi, mu=mu, kappa=kappa)
+            got = pm.logp(dist, 0.0).eval()
+        np.testing.assert_allclose(got, expected, rtol=1e-6)
+
+    def test_logp_at_one(self):
+        zoi, coi, mu, kappa = 0.3, 0.4, 0.4, 10.0
+        expected = np.log(zoi) + np.log(coi)
+        with pm.Model():
+            dist = pm.ZeroOneInflatedBeta.dist(zoi=zoi, coi=coi, mu=mu, kappa=kappa)
+            got = pm.logp(dist, 1.0).eval()
+        np.testing.assert_allclose(got, expected, rtol=1e-6)
+
+    def test_logp_interior(self):
+        zoi, coi, mu, kappa = 0.3, 0.4, 0.4, 10.0
+        alpha = mu * kappa
+        beta = (1 - mu) * kappa
+        for v in [0.1, 0.3, 0.5, 0.7, 0.9]:
+            expected = np.log1p(-zoi) + st.beta.logpdf(v, alpha, beta)
+            with pm.Model():
+                dist = pm.ZeroOneInflatedBeta.dist(zoi=zoi, coi=coi, mu=mu, kappa=kappa)
+                got = pm.logp(dist, v).eval()
+            np.testing.assert_allclose(got, expected, rtol=1e-6, err_msg=f"Mismatch at value={v}")
+
+    def test_support_point(self):
+        zoi, coi, mu, kappa = 0.2, 0.3, 0.4, 10.0
+        expected = zoi * coi + (1 - zoi) * mu
+        with pm.Model():
+            dist = pm.ZeroOneInflatedBeta.dist(zoi=zoi, coi=coi, mu=mu, kappa=kappa)
+            alpha = mu * kappa
+            beta = (1 - mu) * kappa
+            sp = pm.ZeroOneInflatedBeta.support_point(dist, None, zoi, coi, alpha, beta)
+        np.testing.assert_allclose(float(sp), expected, rtol=1e-6)
+
+    def test_parameter_constraints(self):
+        from pymc.logprob.utils import ParameterValueError
+
+        with pytest.raises(ParameterValueError):
+            with pm.Model():
+                dist = pm.ZeroOneInflatedBeta.dist(zoi=1.5, coi=0.4, mu=0.5, kappa=10)
+                pm.logp(dist, 0.5).eval()
+        with pytest.raises(ParameterValueError):
+            with pm.Model():
+                dist = pm.ZeroOneInflatedBeta.dist(zoi=0.3, coi=1.5, mu=0.5, kappa=10)
+                pm.logp(dist, 0.5).eval()
+
+        with pytest.raises(ParameterValueError):
+            with pm.Model():
+                dist = pm.ZeroOneInflatedBeta.dist(zoi=0.3, coi=0.4, alpha=-1.0, beta=5.0)
+                pm.logp(dist, 0.5).eval()
+
+    def test_zib_special_case(self):
+        with pm.Model():
+            y = pm.ZeroOneInflatedBeta("y", zoi=0.5, coi=0.0, mu=0.5, kappa=10, size=1000)
+            prior = pm.sample_prior_predictive(draws=1, random_seed=42)
+        samples = prior.prior["y"].values.flatten()
+        assert np.sum(samples == 1.0) == 0
+
+    def test_oib_special_case(self):
+        with pm.Model():
+            y = pm.ZeroOneInflatedBeta("y", zoi=0.5, coi=1.0, mu=0.5, kappa=10, size=1000)
+            prior = pm.sample_prior_predictive(draws=1, random_seed=42)
+        samples = prior.prior["y"].values.flatten()
+        assert np.sum(samples == 0.0) == 0
+
+    def test_high_zoi_produces_many_boundaries(self):
+        with pm.Model():
+            y = pm.ZeroOneInflatedBeta("y", zoi=0.9, coi=0.5, mu=0.5, kappa=10, size=2000)
+            prior = pm.sample_prior_predictive(draws=1, random_seed=42)
+        samples = prior.prior["y"].values.flatten()
+        prop_boundary = np.mean((samples == 0.0) | (samples == 1.0))
+        assert 0.83 < prop_boundary < 0.95
+
+    def test_zoib_both_boundaries(self):
+        with pm.Model():
+            y = pm.ZeroOneInflatedBeta("y", zoi=0.5, coi=0.5, mu=0.5, kappa=10, size=5000)
+            prior = pm.sample_prior_predictive(draws=1, random_seed=42)
+        samples = prior.prior["y"].values.flatten()
+        prop_zeros = np.mean(samples == 0.0)
+        prop_ones = np.mean(samples == 1.0)
+        prop_interior = np.mean((samples > 0.0) & (samples < 1.0))
+        assert 0.18 < prop_zeros < 0.32
+        assert 0.18 < prop_ones < 0.32
+        assert 0.43 < prop_interior < 0.57
+
+    def test_zoib_reduces_to_beta(self):
+        mu, kappa = 0.4, 10.0
+        alpha = mu * kappa
+        beta = (1 - mu) * kappa
+        for v in [0.2, 0.5, 0.8]:
+            expected = st.beta.logpdf(v, alpha, beta)
+            with pm.Model():
+                dist = pm.ZeroOneInflatedBeta.dist(zoi=0.0, coi=0.5, mu=mu, kappa=kappa)
+                got = pm.logp(dist, v).eval()
+            np.testing.assert_allclose(got, expected, rtol=1e-6, err_msg=f"Mismatch at value={v}")
+
+    def test_zoib_mu_kappa_matches_alpha_beta(self):
+        mu, kappa = 0.4, 10.0
+        alpha = mu * kappa
+        beta = (1 - mu) * kappa
+        zoi, coi = 0.3, 0.4
+        for v in [0.0, 0.5, 1.0]:
+            with pm.Model():
+                dist_mk = pm.ZeroOneInflatedBeta.dist(zoi=zoi, coi=coi, mu=mu, kappa=kappa)
+                logp_mk = pm.logp(dist_mk, v).eval()
+            with pm.Model():
+                dist_ab = pm.ZeroOneInflatedBeta.dist(zoi=zoi, coi=coi, alpha=alpha, beta=beta)
+                logp_ab = pm.logp(dist_ab, v).eval()
+            np.testing.assert_allclose(
+                logp_mk, logp_ab, rtol=1e-6, err_msg=f"Mismatch at value={v}"
+            )
+
+    def test_zoib_vectorized(self):
+        zoi = np.array([0.1, 0.2, 0.3])
+        coi = np.array([0.3, 0.4, 0.5])
+        mu = np.array([0.3, 0.5, 0.7])
+        kappa = np.array([10.0, 15.0, 20.0])
+        values = np.array([0.0, 0.5, 1.0])
+        with pm.Model():
+            dist = pm.ZeroOneInflatedBeta.dist(zoi=zoi, coi=coi, mu=mu, kappa=kappa)
+            logp = pm.logp(dist, values).eval()
+        assert logp.shape == (3,)
+        assert np.all(np.isfinite(logp))
+
+    def test_sampling_recovers_parameters(self):
+        rng = np.random.default_rng(42)
+        true_zoi, true_coi = 0.25, 0.40
+        true_mu, true_kappa = 0.45, 15.0
+        n = 600
+
+        alpha = true_mu * true_kappa
+        beta_p = (1 - true_mu) * true_kappa
+        u = rng.random(n)
+        data = np.where(
+            u < true_zoi * (1 - true_coi),
+            0.0,
+            np.where(u < true_zoi, 1.0, rng.beta(alpha, beta_p, n)),
+        )
+
+        with pm.Model():
+            zoi = pm.Beta("zoi", alpha=1, beta=1)
+            coi = pm.Beta("coi", alpha=1, beta=1)
+            mu = pm.Beta("mu", alpha=2, beta=2)
+            kappa = pm.Gamma("kappa", alpha=2, beta=0.1)
+            pm.ZeroOneInflatedBeta("y", zoi=zoi, coi=coi, mu=mu, kappa=kappa, observed=data)
+            trace = pm.sample(
+                1000,
+                tune=1000,
+                chains=2,
+                random_seed=42,
+                progressbar=False,
+                return_inferencedata=True,
+            )
+
+        post = trace.posterior
+        assert abs(post["zoi"].mean().values - true_zoi) < 0.08
+        assert abs(post["coi"].mean().values - true_coi) < 0.08
+        assert abs(post["mu"].mean().values - true_mu) < 0.08
+        assert abs(post["kappa"].mean().values - true_kappa) < 5.0
+
+    def test_zero_one_inflated_beta_logp(self):
+        def zoib_logp(value, zoi, coi, alpha, beta):
+            if value == 0:
+                return np.log(zoi) + np.log1p(-coi)
+            elif value == 1:
+                return np.log(zoi) + np.log(coi)
+            else:
+                return np.log1p(-zoi) + st.beta.logpdf(value, alpha, beta)
+
+        def zoib_logcdf(value, zoi, coi, alpha, beta):
+            if value < 0:
+                return -np.inf
+            elif value == 0:
+                return np.log(zoi * (1 - coi))
+            elif value < 1:
+                return np.log(zoi * (1 - coi) + (1 - zoi) * st.beta.cdf(value, alpha, beta))
+            else:
+                return 0.0
+
+        check_logp(
+            pm.ZeroOneInflatedBeta,
+            Unit,
+            {"zoi": Unit, "coi": Unit, "alpha": Rplus, "beta": Rplus},
+            zoib_logp,
+        )
+        check_logcdf(
+            pm.ZeroOneInflatedBeta,
+            Unit,
+            {"zoi": Unit, "coi": Unit, "alpha": Rplus, "beta": Rplus},
+            zoib_logcdf,
+        )
